+++ src/admin/admin.module.ts
@@ -8,15 +8,18 @@ import { Invite } from './invite/entity/invite.entity';
 import { JwtModule } from '@nestjs/jwt';
 import authConfig from 'src/config/auth.config';
 import { MailModule } from 'src/mail/mail.module';
+import { WorkspaceModule } from 'src/workspace/workspace.module';
+import { WorkspaceInvite } from './invite/entity/workspace-invite.entity';
 
 @Module({
   controllers: [AdminController],
   providers: [AdminService],
   imports: [
-    TypeOrmModule.forFeature([User, Invite]),
+    TypeOrmModule.forFeature([User, Invite,WorkspaceInvite]),
     UserModule,
     JwtModule.registerAsync(authConfig.asProvider()),
-    MailModule
+    MailModule,
+    WorkspaceModule
   ],
   exports: [AdminService]
 })
