+++ src/admin/admin.service.ts
@@ -1,4 +1,3 @@
-
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreateUserDto } from 'src/user/dtos/create-user.dto';
@@ -10,6 +9,9 @@ import { CreateInviteDto } from './invite/dto/invite.dto';
 import { UserService } from 'src/user/user.service';
 import { randomBytes } from 'crypto';
 import { MailService } from 'src/mail/mail.service';
+import { WorkspaceService } from 'src/workspace/workspace.service';
+import { WorkspaceInvite } from './invite/entity/workspace-invite.entity';
+import { CreateWorkspaceInviteDto } from './invite/dto/ceate-workspace-invite.dto';
 
 @Injectable()
 export class AdminService {
@@ -23,7 +25,11 @@ export class AdminService {
 
         private readonly userService: UserService,
 
-        private readonly mailService: MailService
+        private readonly mailService: MailService,
+
+        @InjectRepository(WorkspaceInvite)
+                private workspaceInviteRepository: Repository<WorkspaceInvite>,
+                private workspaceService: WorkspaceService
     ) { }
 
     async createDefaultAdmin() {
@@ -99,6 +105,53 @@ export class AdminService {
             throw error
         }
     }
+
+     async sendWorkspaceInvite(inviteDto: CreateWorkspaceInviteDto) {
+            try {
+                const { email, workspaceId } = inviteDto;
+    
+                const existingUser = await this.userService.findUserByMail(email)
+    
+                if (!existingUser) {
+                    throw new BadRequestException(`User with email ${email} does not exist`)
+                }
+    
+                const workspace = await this.workspaceService.findWorkspaceById(workspaceId)
+    
+                if (!workspace) {
+                    throw new BadRequestException(`Workspace not found`)
+                }
+    
+                const existingInvite = await this.workspaceInviteRepository.findOne({
+                    where: { email, workspaceId }
+                })
+                if (existingInvite && existingInvite.expiresAt > new Date()) {
+                    throw new BadRequestException(`Invite already sent to email ${email}`)
+                }
+    
+                const token = randomBytes(32).toString('hex')
+                const expiresAt = new Date(Date.now() + 24 * 60 * 60 * 1000)
+    
+                const invite = this.workspaceInviteRepository.create({
+                    email,
+                    token,
+                    workspaceId,
+                    isAccepted: false,
+                    expiresAt
+                })
+    
+                await this.workspaceInviteRepository.save(invite)
+                await this.mailService.sendWorkspaceInvite(email, token)
+    
+                return {
+                    success: true,
+                    message: `Workspace invitation mail sent successfully!`
+                }
+            } catch (error) {
+                throw error
+            }
+        }
+    
 }
